# backend/backend-old/openapi.yaml
openapi: 3.0.0
info:
  title: Pet Adoption API
  description: |
    API for pet adoption web application with user management, animal listings, and messaging.
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. To access protected endpoints:
    1. Get a token by authenticating via `/users/login` or `/admin/login`
    2. Include the token in the Authorization header: `Authorization: Bearer YOUR_TOKEN`
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /admin/login:
    post:
      summary: Authenticate admin
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
        '401':
          description: Invalid credentials
...
  /animals/all:
    get:
      summary: Get all animals with multimedia
      tags:
        - Animals
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all animals with multimedia
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnimalWithMultimedia'
        '401':
          description: Unauthorized
  /animals/details:
    post:
      summary: Get animal details by ID
      tags:
        - Animals
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - animalId
              properties:
                animalId:
                  type: integer
      responses:
        '200':
          description: Detailed animal information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimalDetails'
        '400':
          description: Animal ID is required
        '404':
          description: Animal not found
  /animals/species:
    post:
      summary: Find animals by species
      tags:
        - Animals
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - species
              properties:
                species:
                  type: string
      responses:
        '200':
          description: List of animals of specified species
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnimalWithMultimedia'
        '400':
          description: Species is required
        '404':
          description: No animals found for this species
  /animals/create:
    post:
      summary: Create a new animal
      tags:
        - Animals
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimalCreate'
      responses:
        '201':
          description: Animal created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  animalId:
                    type: integer
        '400':
          description: Missing required fields or invalid data
...
  /animals/top-by-city:
    get:
      summary: Get top animals by city
      tags:
        - Animals
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
          description: User ID to determine the city
      responses:
        '200':
          description: List of top animals in user's city
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Animal'
        '400':
          description: User ID is required
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Authentication using Bearer tokens.
        Get your token by authenticating through the /users/login or /admin/login endpoints.
        
        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
    User:
      type: object
      properties:
        USERID:
          type: integer
        FIRSTNAME:
          type: string
        LASTNAME:
          type: string
        EMAIL:
          type: string
          format: email
        PHONE:
          type: string
        CREATEDAT:
          type: string
          format: date-time
    Address:
      type: object
      properties:
        ADDRESSID:
          type: integer
        STREET:
          type: string
        CITY:
          type: string
        STATE:
          type: string
        ZIPCODE:
          type: string
        COUNTRY:
          type: string
...
    Animal:
      type: object
      properties:
        ANIMALID:
          type: integer
        USERID:
          type: integer
        NAME:
          type: string
        BREED:
          type: string
        SPECIES:
          type: string
        AGE:
          type: integer
        GENDER:
          type: string
          enum:
            - male
            - female
        VIEWS:
          type: integer
        CREATEDAT:
          type: string
          format: date-time
    MultiMedia:
      type: object
      properties:
        ID:
          type: integer
        ANIMALID:
          type: integer
        MEDIATYPE:
          type: string
        URL:
          type: string
        DESCRIPTION:
          type: string
        UPLOAD_DATE:
          type: string
          format: date-time
    AnimalWithMultimedia:
      allOf:
        - $ref: '#/components/schemas/Animal'
        - type: object
          properties:
            multimedia:
              type: array
              items:
                $ref: '#/components/schemas/MultiMedia'
    FeedingSchedule:
      type: object
      properties:
        ID:
          type: integer
        ANIMALID:
          type: integer
        FEEDING_TIME:
          type: array
          items:
            type: string
            format: time
        FOOD_TYPE:
          type: string
        NOTES:
          type: string
    MedicalHistory:
      type: object
      properties:
        ID:
          type: integer
        ANIMALID:
          type: integer
        VETNUMBER:
          type: string
        RECORDDATE:
          type: string
          format: date
        DESCRIPTION:
          type: string
        FIRST_AID_NOTED:
          type: boolean
    Relations:
      type: object
      properties:
        ID:
          type: integer
        ANIMALID:
          type: integer
        FRIENDWITH:
          type: string
    AnimalDetails:
      type: object
      properties:
        animal:
          $ref: '#/components/schemas/Animal'
        multimedia:
          type: array
          items:
            $ref: '#/components/schemas/MultiMedia'
        feedingSchedule:
          type: array
          items:
            $ref: '#/components/schemas/FeedingSchedule'
        medicalHistory:
          type: array
          items:
            $ref: '#/components/schemas/MedicalHistory'
        owner:
          $ref: '#/components/schemas/User'
        address:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        relations:
          type: array
          items:
            $ref: '#/components/schemas/Relations'
...
    AnimalCreate:
      type: object
      required:
        - userID
        - name
        - breed
        - species
        - age
        - gender
      properties:
        userID:
          type: integer
        name:
          type: string
        breed:
          type: string
        species:
          type: string
        age:
          type: integer
        gender:
          type: string
          enum:
            - male
            - female
        feedingSchedule:
          type: array
          items:
            type: object
            properties:
              feedingTime:
                type: string
                format: time
              foodType:
                type: string
        medicalHistory:
          type: array
          items:
            type: object
            properties:
              vetNumber:
                type: string
              recordDate:
                type: string
                format: date
              description:
                type: string
              first_aid_noted:
                type: boolean
        multimedia:
          type: array
          items:
            type: object
            properties:
              mediaType:
                type: string
              url:
                type: string
              description:
                type: string
        relations:
          type: object
          properties:
            friendWith:
              type: string
...